Moved module test to another file
